Start

+++ src/core/auth/services/auth.service.ts
@@ -1,20 +1,20 @@
-import { TYPES } from "@DI/types";
-import { HttpStatus, Inject, Injectable } from "@nestjs/common";
-import { HttpException } from "@shared/exceptions";
-import { IBcryptService } from "@src/common/auth/bcrypt";
-import { IJWTService } from "@src/common/auth/jwt";
-import { IAuthRepository } from "@src/core/auth";
-import { IUserRepository, IUserService } from "@src/core/user";
-import { CreateUserDto } from "@src/core/user/dto";
-import { User } from "@src/core/user/types";
+import { TYPES } from '@DI/types';
+import { HttpStatus, Inject, Injectable } from '@nestjs/common';
+import { HttpException } from '@shared/exceptions';
+import { IBcryptService } from '@src/common/auth/bcrypt';
+import { IJWTService } from '@src/common/auth/jwt';
+import { IAuthRepository } from '@src/core/auth';
+import { IUserRepository, IUserService } from '@src/core/user';
+import { CreateUserDto, ResultUserDto } from '@src/core/user/dto';
 
-import { IAuthCacheService } from "../cache";
-import { UserLoginDto } from "../dto";
+import { IAuthCacheService } from '../cache';
+import { UserLoginDto } from '../dto';
 
 export interface IAuthService {
   login: (loginDto: UserLoginDto) => Promise<{ accessToken: string; refreshToken: string }>;
   register: (registerDto: CreateUserDto) => Promise<{ id: string } | never>;
   refresh: (token: string) => Promise<{ accessToken: string; refreshToken: string } | never>;
+  logout: (userId: string) => Promise<boolean>;
 }
 
 @Injectable()
@@ -30,11 +30,6 @@ export class AuthService implements IAuthService {
 
   async register({ name, password, role }: CreateUserDto) {
     return this.userService.create({ name, password, role });
-    // return this.authRepository.register({
-    //   name,
-    //   password: await this.bcryptService.hash(password),
-    //   role: [],
-    // });
   }
 
   async refresh(refreshToken: string) {
@@ -45,14 +40,14 @@ export class AuthService implements IAuthService {
       data: string;
     };
 
-    if (!token) throw new HttpException("Invalid refresh token", HttpStatus.BAD_REQUEST);
+    if (!token) throw new HttpException('Invalid refresh token', HttpStatus.BAD_REQUEST);
 
     const tokenFromCache = await this.authCacheService.getTokenFromCache(
-      this.authCacheService.getTokenCacheName("refresh", token.data),
+      this.authCacheService.getTokenCacheName('refresh', token.data),
     );
 
     if (tokenFromCache !== refreshToken)
-      throw new HttpException("Invalid refresh token", HttpStatus.BAD_REQUEST);
+      throw new HttpException('Invalid refresh token', HttpStatus.BAD_REQUEST);
 
     const tokens = this.jwtService.generateTokens(token.data, token.data);
 
@@ -65,7 +60,7 @@ export class AuthService implements IAuthService {
     return tokens;
   }
 
-  private async verifyUser(name: string, password: string): Promise<false | User> {
+  private async verifyUser(name: string, password: string): Promise<false | ResultUserDto> {
     const user = await this.userService.findByName(name);
 
     if (!user) return false;
@@ -79,11 +74,11 @@ export class AuthService implements IAuthService {
     const user = await this.verifyUser(name, password);
 
     if (!user) {
-      throw new HttpException("Не верный логин или пароль", HttpStatus.BAD_REQUEST);
+      throw new HttpException('Не верный логин или пароль', HttpStatus.BAD_REQUEST);
     }
 
     if (!user.settings.active) {
-      throw new HttpException("Пользователь не активен", HttpStatus.BAD_REQUEST);
+      throw new HttpException('Пользователь не активен', HttpStatus.BAD_REQUEST);
     }
 
     const tokens = this.jwtService.generateTokens(user.id, user.id);
@@ -94,25 +89,9 @@ export class AuthService implements IAuthService {
   }
 
   public async logout(userId: string): Promise<boolean> {
+    if (!userId) {
+      throw new HttpException('Ошибка выхода из аккаунта', HttpStatus.BAD_REQUEST);
+    }
     return this.authCacheService.removeTokens(userId);
   }
-
-  // async login({ email, password }: UserLoginDto) {
-  //   const user = await this.userRepository.findUser({ name: email });
-  //
-  //   if (!user) {
-  //     throw new HttpException(`User with email ${email} not found`, 404);
-  //   }
-  //
-  //   const isPasswordValid = await this.bcryptService.compare(password, user.password);
-  //
-  //   if (!isPasswordValid) {
-  //     throw new HttpException("Incorrect password", HttpStatus.BAD_REQUEST);
-  //   }
-  //   const tokens = this.jwtService.generateTokens(user.id, user.id);
-  //
-  //   await this.authCacheService.setTokensToCache(user.id, tokens.accessToken, tokens.refreshToken);
-  //
-  //   return tokens;
-  // }
 }
